socketserver/pyproject.toml
[tool.poetry]
name = "clintochatserver"
version = "0.1.0"
description = ""
authors = ["sakthiRatnam <sakthiratnam050@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-socketio = "^5.5.0"
gunicorn = "^20.1.0"
eventlet = "^0.33.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
